requirements.txt
web3
python-dotenv
requests
flask
eth-abi
